% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{go2GRangesList}
\alias{go2GRangesList}
\title{go2GRangesList}
\usage{
go2GRangesList(goTerms, txdb = "TxDb.Hsapiens.UCSC.hg19.knownGene",
  searchDist = 0, parallel = T)
}
\description{
From a vector of GO terms, create a GRangesList
holding TSS +/- a window of variable length.
}
