% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/code.R
\name{matchAnnotations}
\alias{matchAnnotations}
\title{matchAnnotations}
\usage{
matchAnnotations(annotations, rse)
}
\description{
From a RangedSummarizedExperiment, create a new
RangedSummarizedExperiment with a sparseMatrix in the assay
slot.
}
